+++ src/refactoring/hooks/useCoupon.ts
@@ -1,6 +1,5 @@
 import { useState } from "react";
 import { Coupon } from "../../types.ts";
-import { initialCoupons } from "../../constants/index.ts";
 
 
 export const useCoupons = (initialCoupons: Coupon[]) => {
